requirements.txt
agno[ddg,website,pandas,csv,file,calculator,email,yfinance,duckdb]
psycopg2-binary
sqlalchemy
pgvector
python-dotenv
openai
azure-identity
pypdf
ddgs
exa-py
slack-sdk
yfinance
pandas
requests
beautifulsoup4
lxml
newspaper4k
duckdb 